+++ src/richie/apps/search/__init__.py
@@ -1,9 +1,16 @@
 """Define SearchConfig as the default app configuration."""
 from django.conf import settings
 
-from elasticsearch import Elasticsearch
+from .elasticsearch import (
+    ElasticsearchClientCompat7to6,
+    ElasticsearchIndicesClientCompat7to6,
+)
 
 # pylint: disable=invalid-name
 default_app_config = "richie.apps.search.apps.SearchConfig"
 
-ES_CLIENT = Elasticsearch([getattr(settings, "RICHIE_ES_HOST", "elasticsearch")])
+ES_CLIENT = ElasticsearchClientCompat7to6(
+    [getattr(settings, "RICHIE_ES_HOST", "elasticsearch")]
+)
+
+ES_INDICES_CLIENT = ElasticsearchIndicesClientCompat7to6(ES_CLIENT)