+++ tests/apps/search/test_autocomplete_persons.py
@@ -6,10 +6,8 @@ from unittest import mock
 
 from django.test import TestCase
 
-from elasticsearch.client import IndicesClient
-from elasticsearch.helpers import bulk
-
-from richie.apps.search import ES_CLIENT
+from richie.apps.search import ES_CLIENT, ES_INDICES_CLIENT
+from richie.apps.search.elasticsearch import bulk_compat
 from richie.apps.search.indexers.persons import PersonsIndexer
 from richie.apps.search.text_indexing import ANALYSIS_SETTINGS
 from richie.apps.search.utils.indexers import slice_string_for_completion
@@ -54,21 +52,18 @@ class AutocompletePersonsTestCase(TestCase):
             },
         ]
 
-        indices_client = IndicesClient(client=ES_CLIENT)
         # Delete any existing indices so we get a clean slate
-        indices_client.delete(index="_all")
+        ES_INDICES_CLIENT.delete(index="_all")
         # Create an index we'll use to test the ES features
-        indices_client.create(index=PERSONS_INDEX)
+        ES_INDICES_CLIENT.create(index=PERSONS_INDEX)
 
         # The index needs to be closed before we set an analyzer
-        indices_client.close(index=PERSONS_INDEX)
-        indices_client.put_settings(body=ANALYSIS_SETTINGS, index=PERSONS_INDEX)
-        indices_client.open(index=PERSONS_INDEX)
+        ES_INDICES_CLIENT.close(index=PERSONS_INDEX)
+        ES_INDICES_CLIENT.put_settings(body=ANALYSIS_SETTINGS, index=PERSONS_INDEX)
+        ES_INDICES_CLIENT.open(index=PERSONS_INDEX)
 
         # Use the default persons mapping from the Indexer
-        indices_client.put_mapping(
-            body=PersonsIndexer.mapping, index=PERSONS_INDEX
-        )
+        ES_INDICES_CLIENT.put_mapping(body=PersonsIndexer.mapping, index=PERSONS_INDEX)
 
         # Actually insert our persons in the index
         actions = [
@@ -82,8 +77,8 @@ class AutocompletePersonsTestCase(TestCase):
             }
             for person in persons
         ]
-        bulk(actions=actions, chunk_size=500, client=ES_CLIENT)
-        indices_client.refresh()
+        bulk_compat(actions=actions, chunk_size=500, client=ES_CLIENT)
+        ES_INDICES_CLIENT.refresh()
 
         response = self.client.get(
             f"/api/v1.0/persons/autocomplete/?{querystring:s}", **extra